+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Devel
 
 * Bugfix: `run_anvil_audit` now raises a `CommandException` if an API call fails.
+* Bugfix: use the workspace form specified in the workspace adapter when cloning a workspace.
 
 ## 0.26.0 (2024-11-08)